chore: split out function types into multiple files
