+++ src/index.ts
@@ -31,7 +31,7 @@ export * from './types/blueprints.js'
 export * from './types/cors.js'
 export * from './types/datasets.js'
 export * from './types/errors.js'
-export * from './types/functions.js'
+export * from './types/functions/index.js'
 export * from './types/projects.js'
 export * from './types/resources.js'
 export * from './types/robots.js'
